Feat: add text class for blocks of text
