+++ mastermind.rb
@@ -1,4 +1,5 @@
 require './lib/game'
+require './lib/text'
 
 # puts "Welcome to MASTERMIND"
 # puts "Would you like to (p)lay, read the (i)nstructions, or (q)uit?"
@@ -7,17 +8,17 @@ require './lib/game'
 
 
 game = Game.new
+text = Text.new
 
 puts "Welcome to MASTERMIND"
 
 loop do
-  puts "Would you like to (p)lay, read the (i)nstructions, or (q)uit?"
-  print ">"
+  text.welcome
   input = gets.chomp.downcase
   if input == 'p' || input == 'play'
     game.start
   elsif input == 'i' || input == 'instructions'
-    game.instructions # game.insructions
+    text.instructions # game.insructions
   elsif input == 'q' || input == 'quit'
     break
   end
